requirements.txt
flask
pymongo
python-dotenv
Requests
selenium
webdriver-manager

